+++ src/main/java/refinedstorage/container/ContainerSolderer.java
@@ -20,13 +20,13 @@ public class ContainerSolderer extends ContainerBase {
             y += 18;
         }
 
-        addSlotToContainer(new SlotOutput(solderer.getItems(), 3, 134, 38));
+        addSlotToContainer(new SlotOutput(solderer.getItems(), 3, 127, 38));
 
         for (int i = 0; i < 4; ++i) {
             addSlotToContainer(new SlotItemHandler(solderer.getUpgrades(), i, 187, 6 + (i * 18)));
         }
 
-        addPlayerInventory(8, 95);
+        addPlayerInventory(8, 89);
     }
 
     @Override