+++ src/main/java/refinedstorage/gui/GuiSolderer.java
@@ -8,7 +8,7 @@ public class GuiSolderer extends GuiBase {
     private TileSolderer solderer;
 
     public GuiSolderer(ContainerSolderer container, TileSolderer solderer) {
-        super(container, 211, 177);
+        super(container, 211, 171);
 
         this.solderer = solderer;
     }
@@ -29,13 +29,17 @@ public class GuiSolderer extends GuiBase {
         drawTexture(x, y, 0, 0, width, height);
 
         if (solderer.isWorking()) {
-            drawTexture(x + 83, y + 40 - 1, 212, 0, solderer.getProgressScaled(22), 15);
+            drawTexture(x + 83, y + 38 - 1, 212, 0, solderer.getProgressScaled(22), 15);
         }
     }
 
     @Override
     public void drawForeground(int mouseX, int mouseY) {
         drawString(7, 7, t("gui.refinedstorage:solderer"));
-        drawString(7, 82, t("container.inventory"));
+        drawString(7, 77, t("container.inventory"));
+
+        if (inBounds(83, 38, 22, 15, mouseX, mouseY) && solderer.isWorking()) {
+            drawTooltip(mouseX, mouseY, t("gui.refinedstorage:solderer.progress", solderer.getProgressScaled(100)));
+        }
     }
 }